Remove redundant annotations; Refactor package structure;
Closes #22
