+++ src/main/java/ru/aahzbrut/reciperestapi/domain/Difficulty.java
@@ -1,4 +1,4 @@
-package ru.aahzbrut.reciperestapi.unit.entities;
+package ru.aahzbrut.reciperestapi.domain;
 
 public enum Difficulty {
     EASY, MEDIUM, HARD, IMPOSSIBLE
