+++ src/main/java/ru/aahzbrut/reciperestapi/domain/entities/Recipe.java
@@ -1,4 +1,4 @@
-package ru.aahzbrut.reciperestapi.unit.entities;
+package ru.aahzbrut.reciperestapi.domain.entities;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -6,7 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import ru.aahzbrut.reciperestapi.unit.entities.domain.BaseEntity;
+import ru.aahzbrut.reciperestapi.domain.BaseEntity;
+import ru.aahzbrut.reciperestapi.domain.Difficulty;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.CascadeType;
