+++ src/main/java/ru/aahzbrut/reciperestapi/domain/entities/UOM.java
@@ -1,21 +1,17 @@
-package ru.aahzbrut.reciperestapi.unit.entities;
+package ru.aahzbrut.reciperestapi.domain.entities;
 
-import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import ru.aahzbrut.reciperestapi.unit.entities.domain.BaseEntity;
+import ru.aahzbrut.reciperestapi.domain.BaseEntity;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.ManyToMany;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
-import java.util.LinkedList;
-import java.util.List;
 
 @Entity
 @Getter
@@ -23,10 +19,10 @@ import java.util.List;
 @NoArgsConstructor
 @SuperBuilder
 @EqualsAndHashCode(callSuper = true, onlyExplicitlyIncluded = true)
-@SequenceGenerator(name = "default_gen", sequenceName = "CATEGORY_SEQ", allocationSize = 1)
-@AttributeOverride(name = "id", column = @Column(name = "CATEGORY_ID"))
-@Table(name = "CATEGORY")
-public class Category extends BaseEntity {
+@SequenceGenerator(name = "default_gen", sequenceName = "UOM_SEQ", allocationSize = 1)
+@AttributeOverride(name = "id", column = @Column(name = "UOM_ID"))
+@Table(name = "UOM")
+public class UOM extends BaseEntity {
 
     private static final long serialVersionUID = 1L;
 
@@ -35,8 +31,4 @@ public class Category extends BaseEntity {
 
     @Column(name = "DESCRIPTION")
     private String description;
-
-    @ManyToMany(mappedBy = "categories")
-    @Builder.Default
-    private List<Recipe> recipes = new LinkedList<>();
 }
