+++ src/main/java/ru/aahzbrut/reciperestapi/repositories/CategoryRepository.java
@@ -2,7 +2,7 @@ package ru.aahzbrut.reciperestapi.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.aahzbrut.reciperestapi.unit.entities.Category;
+import ru.aahzbrut.reciperestapi.domain.entities.Category;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<Category, Long> {
