+++ src/main/java/ru/aahzbrut/reciperestapi/repositories/NoteRepository.java
@@ -2,7 +2,7 @@ package ru.aahzbrut.reciperestapi.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.aahzbrut.reciperestapi.unit.entities.Note;
+import ru.aahzbrut.reciperestapi.domain.entities.Note;
 
 @Repository
 public interface NoteRepository extends JpaRepository<Note, Long> {
