+++ src/main/java/ru/aahzbrut/reciperestapi/repositories/RecipeStepRepository.java
@@ -2,7 +2,7 @@ package ru.aahzbrut.reciperestapi.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.aahzbrut.reciperestapi.unit.entities.RecipeStep;
+import ru.aahzbrut.reciperestapi.domain.entities.RecipeStep;
 
 @Repository
 public interface RecipeStepRepository extends JpaRepository<RecipeStep, Long> {
