+++ src/main/java/ru/aahzbrut/reciperestapi/repositories/StepRepository.java
@@ -2,7 +2,7 @@ package ru.aahzbrut.reciperestapi.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.aahzbrut.reciperestapi.unit.entities.Step;
+import ru.aahzbrut.reciperestapi.domain.entities.Step;
 
 @Repository
 public interface StepRepository extends JpaRepository<Step, Long> {
