+++ src/main/java/ru/aahzbrut/reciperestapi/repositories/UOMRepository.java
@@ -2,7 +2,7 @@ package ru.aahzbrut.reciperestapi.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.aahzbrut.reciperestapi.unit.entities.UOM;
+import ru.aahzbrut.reciperestapi.domain.entities.UOM;
 
 @Repository
 public interface UOMRepository extends JpaRepository<UOM, Long> {
