+++ src/test/java/ru/aahzbrut/reciperestapi/entities/domain/NoteTest.java
@@ -1,15 +1,15 @@
-package ru.aahzbrut.reciperestapi.unit.entities.domain;
+package ru.aahzbrut.reciperestapi.entities.domain;
 
 import org.junit.jupiter.api.Test;
-import ru.aahzbrut.reciperestapi.unit.entities.UOM;
+import ru.aahzbrut.reciperestapi.domain.entities.Note;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class UOMTest {
+class NoteTest {
 
     @Test
     void npeDefence() {
-        TestUtils.npeDefence(UOM.class);
+        TestUtils.npeDefence(Note.class);
         int one = 1;
         assertEquals(1, one);
     }
