+++ src/test/java/ru/aahzbrut/reciperestapi/entities/domain/RecipeTest.java
@@ -1,17 +1,18 @@
-package ru.aahzbrut.reciperestapi.unit.entities.domain;
+package ru.aahzbrut.reciperestapi.entities.domain;
 
 import org.junit.jupiter.api.Test;
-import ru.aahzbrut.reciperestapi.unit.entities.RecipeStep;
+import ru.aahzbrut.reciperestapi.domain.entities.Recipe;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class RecipeStepTest {
+class RecipeTest {
 
     @Test
     void npeDefence() {
-        TestUtils.npeDefence(RecipeStep.class);
+        TestUtils.npeDefence(Recipe.class);
         int one = 1;
         assertEquals(1, one);
     }
+
 
 }
