+++ src/test/java/ru/aahzbrut/reciperestapi/entities/domain/StepIngredientTest.java
@@ -1,7 +1,7 @@
-package ru.aahzbrut.reciperestapi.unit.entities.domain;
+package ru.aahzbrut.reciperestapi.entities.domain;
 
 import org.junit.jupiter.api.Test;
-import ru.aahzbrut.reciperestapi.unit.entities.StepIngredient;
+import ru.aahzbrut.reciperestapi.domain.entities.StepIngredient;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
