+++ src/test/java/ru/aahzbrut/reciperestapi/entities/domain/TestUtils.java
@@ -1,7 +1,8 @@
-package ru.aahzbrut.reciperestapi.unit.entities.domain;
+package ru.aahzbrut.reciperestapi.entities.domain;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
+import ru.aahzbrut.reciperestapi.domain.BaseEntity;
 
 import java.lang.reflect.Method;
 import java.util.stream.Stream;
